requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
Werkzeug==2.0.1
WTForms==2.3.3
email-validator==1.1.3
python-slugify==5.0.2
Pygments==2.10.0
Flask-Migrate==3.1.0
SQLAlchemy==2.0.21
bcrypt==4.0.1
markdown2==2.4.10
